% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_scale.r
\name{col_scale}
\alias{col_scale}
\title{create a color scale legend for a plot}
\usage{
col_scale(
  z,
  zlim,
  col = heat.colors(12),
  breaks,
  horiz = TRUE,
  ylim = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{z}{numeric vector, the input vector}
}
\value{
a color scale legend
}
\description{
create a color scale legend for a plot
}
