% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTVSpectralQuant.r
\name{computeTVSpectralQuant}
\alias{computeTVSpectralQuant}
\title{Compute Time-Varying Spectral Quantities}
\usage{
computeTVSpectralQuant(
  X,
  id,
  windowLen = windowLen,
  normalize = FALSE,
  overlap = 0,
  M_welch = round_right(windowLen/20),
  logScale = FALSE,
  padEnd = TRUE,
  transform = "FZ",
  plot = FALSE,
  bands_analysis = FALSE,
  f_sampling = NULL
)
}
\arguments{
\item{X}{numeric array, Data.}

\item{id}{Numeric Identifier:
- If id is less than or equal to the number of columns (channels) of X, it computes the
auto-spectrum of the id-th channel.
- If id is greater than the number of columns, it computes the cross-coherence between
the two channels corresponding to the id.}

\item{windowLen}{numeric Number of observations used for the estimation of localized spectral quantities.}

\item{normalize}{logical Indicates whether to normalize the spectral quantities.}

\item{overlap}{numeric Amount of overlap between successive windows.}

\item{M_welch}{numeric, Welch regularisation parameter.}

\item{logScale}{logical Indicates whether to use a logarithmic scale for the spectral quantities.}

\item{padEnd}{logical Specifies whether to pad time series with zeros at the end if the number of
observations is not a multiple of windowLen. Default is TRUE.}

\item{transform}{character Transformation to apply to the data.}

\item{plot}{logical Indicates whether to plot the resulting spectral quantities.}

\item{bands_analysis}{logical Indicates whether to analyze by frequency bands.}

\item{f_sampling}{numeric Sampling frequency.}
}
\value{
array Depending on the id:
- If id corresponds to a single channel, it returns the auto-spectrum of that channel.
- If id corresponds to a pair of channels, it returns the cross-coherence between them.
}
\description{
This function calculates spectral quantities either for an individual channel or cross-coherence
between a pair of channels.
}
