% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_stat.r
\name{cusum_stat}
\alias{cusum_stat}
\title{CUSUM statistic}
\usage{
cusum_stat(
  S,
  thresh = 0.8 * log(dim(S)[2])^(1.1),
  check_neighborhood = TRUE,
  check_par = floor(log(dim(S)[2])^2.1/5),
  cp = array(0, dim(S)),
  s = 1,
  delta = 1
)
}
\arguments{
\item{S}{numeric array, The estimated spectral quantities for each window (lines) and for the corresponding frequencies (columns)}

\item{thresh}{numeric, The threshold to decide that a point is a change point, using the statistics eq (5), the 
(default is the theoretical threshold \tau_T Theorem 1 Shorder & Ombao 2019)}

\item{check_neighborhood}{logical, Within the change-point detection, should the neighbourhood around 
a changepoint candidate be tested to reduce the risk of spurious detections?}

\item{check_par}{numeric, \Delta_T the diameter of the interval around a change-point used in check_neighborhood}

\item{cp}{numeric array, containing only seros and has the same dimension as S}

\item{s}{numeric, The first index of the interval for which FreSpeD is applied}

\item{delta}{numeric, \delta_T' in the algorithm: the minimal distance between two change points}
}
\value{
numric array, the values of UHcusum on the detected change points for each frquencies, and 0 otherwise
}
\description{
This function applies the UHcusum function to the estimated spectral quantities. If the cusum value is bigger than
the parameter thresh, the block is considered as a change point.
}
