% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_identifier.r
\name{define_identifier}
\alias{define_identifier}
\title{Define identifiers for EEG channels}
\usage{
define_identifier(D)
}
\arguments{
\item{D}{Numeric, number of the EEG channels}
}
\value{
Numeric vector, representing identifiers for EEG channels
}
\description{
Define identifiers for EEG channels to represent unique pairs of channels.
The identifiers are generated based on the number of EEG channels provided.
}
\examples{
define_identifier(4)  # Returns: 1 2 3 4 102 103 104 203 204 304
}
