% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreSpeD_innerpost.r
\name{FreSpeD_innerpost}
\alias{FreSpeD_innerpost}
\title{Check change point}
\usage{
FreSpeD_innerpost(
  S,
  thresh,
  check_neighborhood = TRUE,
  check_par = floor(log(dim(S)[2])^2.1/5),
  b,
  delta = 1
)
}
\arguments{
\item{S}{numeric array, the estimated spectral quantities}

\item{thresh}{numeric, numeric, The threshold to decide that a point is a change point}

\item{check_neighborhood}{logical, logical, Within the change-point detection, should the neighbourhood around
a changepoint candidate be tested}

\item{b}{numeric, the candidate to be a change point}

\item{delta}{numeric, the minimal distance between two change points}
}
\value{
logical, if b is a spurious change point return FALSE, and TRUE otherwise.
}
\description{
Check wether a given point b is a true change point or spurious change point
}
