% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreSpeD_plot.r
\name{FreSpeD_plot}
\alias{FreSpeD_plot}
\title{Plots the time-varying spectrum and cross-coherence}
\usage{
FreSpeD_plot(
  X,
  cp,
  id,
  channelNames = 1:dim(X)[2],
  windowLen,
  nFB = round_right(windowLen/20),
  overlap = 0,
  normalize = FALSE,
  padEnd = TRUE
)
}
\arguments{
\item{X}{numeric array, Data}

\item{cp}{list of numeric arrays, List of change points per channel/channel pair given by FreSpeD function}

\item{id}{numeric, Which list entry should be plotted?}

\item{channelNames}{numeric vector, channel names. Default is 1:dim(X)\link{2}}

\item{windowLen}{numeric, Number of observations used for the estimation of localized spectral quantities}

\item{nFB}{numeric, Welch regularisation parameter.}

\item{overlap}{numeric, Number of observations by which localized spectral estimates can overlap. Default is zero}

\item{normalize}{logical, Normalize spectrum (dividing by total localized variance)?}

\item{padEnd}{logical, Pad time series with zeros at end (if number of observations in time is not a multiple of windowLen).}
}
\value{
no output is returned
}
\description{
Plots the time-varying spectrum for a specified channel or coherence for a channel pair
along with the estimated change points in time and frequency provided by FreSpeD.
}
