% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreSpeD_postprocessing.r
\name{FreSpeD_postprocessing}
\alias{FreSpeD_postprocessing}
\title{Optional postprocessing}
\usage{
FreSpeD_postprocessing(
  X,
  cp,
  windowLen,
  singleTimeSeries = TRUE,
  allTimeSeries = TRUE,
  clusterradius = c(),
  deltaPer = 0.03,
  nFB = round_right(windowLen/20),
  overlap = 0,
  normalize = FALSE,
  logScale = FALSE,
  padEnd = TRUE,
  transform = "FZ",
  threshFun = function(Tv) {
     0.8 * log(Tv)^(1.1)
 }
)
}
\arguments{
\item{X}{numeric array, Data}

\item{cp}{list of numeric arrays, List of change points per channel/channel pair given by FreSpeD function}

\item{singleTimeSeries}{logical, Check individual channels and channel pairs? This tests every change point of a
given channel/channel pair versus its adjacent change-point neighbors. The test
statistic is consistent with the initial thresholded CUSUM statistic.}

\item{allTimeSeries}{logical, Merge change points over all channels and channel pairs into global clusters.
A cluster is defined as a group of change points where each is at most clusterradius
distant from the nearest cluster member. The resulting merged change point is
the median of all cluster members}

\item{clusterradius}{list Maximum distance allowed between members of the clusters defined through
allTimeSeries(default is empty list)}

\item{overlap}{numeric, the overlap between consecutive windows (default = 0)}

\item{normalize}{logical Indicates whether to normalize the spectral quantities.}

\item{logScale}{logical, Indicates whether to use a logarithmic scale to spectral quantities}

\item{padEnd}{logical, logical, Pad time series with zeros at end (if number of observations in time is not a multiple of windowLen). Default is TRUE}

\item{transform}{character, transformation of the estimate. Fisher-z transform is default}

\item{thresFun}{function, Threshold function}
}
\value{
If only individual tests are conducted, the output is a list similar to that of FreSpeD(). Otherwise,
a single vector of merged change points is provided.
}
\description{
Optional postprocessing to account for information of channel clusters
}
