% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreSpeD_summary.r
\name{FreSpeD_summary}
\alias{FreSpeD_summary}
\title{Summary of FreSpeD output}
\usage{
FreSpeD_summary(cp, channelNames = c(), windowLen = 1, plot = TRUE)
}
\arguments{
\item{cp}{list of numeric arrays, List of change points per channel/channel pair given by FreSpeD function}

\item{channelNames}{list, default empty.}

\item{plot}{logical, Should global summary plots be provided?}

\item{windoLen}{numeric, default 1}
}
\value{
list containing
1. numeric, total number of change points
2. numeric, total number of change points in autospectra
3. numeric vector, change points over time
4. numeric vector, autospectral change points over time
5. numeric vector, change points over frequency bands
6. numeric vector, autospectral change points over frequency bands
7. dataframe, containing the number and location for each channel and channel pair
}
\description{
Various tables and plots showing change point distribution in time, frequency and for individual
channels and channel pairs.
}
