% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bands_indices.r
\name{get_bands_indices}
\alias{get_bands_indices}
\title{Get the indices of the frequency bands}
\usage{
get_bands_indices(length_window, f_sampling)
}
\arguments{
\item{length_window}{Numeric vector, representing the length of the number of equispaced fundamental frequencies in \link{0, 0.5}.}

\item{f_sampling}{Numeric, representing the sampling rate frequency.}
}
\value{
List of vectors, representing the indices of the bands delta, theta, alpha, beta and gamma.
}
\description{
Given the length of the window and the sampling frequency, return the indices of the bands delta, theta, alpha,
beta and gamma.
}
