% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_identifiers.r
\name{name_identifiers}
\alias{name_identifiers}
\title{Name identifiers for EEG channels}
\usage{
name_identifiers(n)
}
\arguments{
\item{n}{Numeric vector, representing EEG channel numbers}
}
\value{
Character vector, representing named identifiers for EEG channels
}
\description{
Generates named identifiers for EEG channels based on unique pairs of channels.

Generates named identifiers for EEG channels based on unique pairs of channels.
}
\examples{
name_identifiers(c(1:4)) # Returns: "1" "2" "3" "4" "1.2" "1.3" "1.4" "2.3" "2.4" "3.4"
}
