% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.r
\name{partition}
\alias{partition}
\title{Partition}
\usage{
partition(X, windowLen, overlap = 0, padEnd = TRUE)
}
\arguments{
\item{X}{Numeric vector, the series to partition}

\item{windowLen}{Numeric, the length of the window}

\item{overlap}{Numeric, the overlap between consecutive windows (default = 0)}

\item{padEnd}{Logical, Pad time series with zeros at end (if number of observations in time is not a
multiple of windowLen). Default is TRUE}
}
\value{
Array, a 2D array where each column contains a partitioned window
}
\description{
Partition the series of numbers into windows, and stack them vertically
}
\examples{
series <- c(1:20)
partitioned <- partition(series, windowLen = 3, overlap = 1)
}
