% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_right.r
\name{round_right}
\alias{round_right}
\title{Round right}
\usage{
round_right(x)
}
\arguments{
\item{x}{Numeric vector, the input vector}
}
\value{
Numeric vector, rounded values
}
\description{
Round to the nearest integer towards positive infinity
}
\examples{
round_right(c(1.1, 2.3, 3.5, 4.7, 5.9))
}
