% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_coherence.r
\name{tv_coherence}
\alias{tv_coherence}
\title{time varying coherence of two signals}
\usage{
tv_coherence(
  X,
  windowLen,
  overlap = 0,
  M_welch = windowLen/2 * 0.1,
  logScale = FALSE,
  padEnd = TRUE,
  transform = "FZ",
  plot = FALSE,
  bands_analysis = FALSE,
  f_sampling = NULL
)
}
\arguments{
\item{X}{numeric two-dimensional data input, rows: observations in time, columns: channels.}

\item{windowLen}{numeric Number of observations used for the estimation of localized spectral quantities.}

\item{overlap}{numeric, the overlap between consecutive windows (default = 0)}

\item{M_welch}{numeric, Welch regularisation parameter.}

\item{logScale}{logical, Indicates whether to use a logarithmic scale to the estimated coherence}

\item{padEnd}{logical, Pad time series with zeros at end (if number of observations in time is not a multiple of windowLen). Default is TRUE}

\item{plot}{logical Indicates whether to plot the estimated cross-coherence}

\item{bands_analysis}{logical, if TRUE, the function will make the analysis by bands}

\item{f_sampling}{numeric, sampling frequency}

\item{tranform}{character, Apply a tranformation to the coherence (default = 'FZ')
'FZ' Fisher-z transformation, 'ARC' arcsin the inverse of sin function, '' no transformation is applied}
}
\value{
list, Output a list containing the estimated cross-coherence, the set of windows used in the estimation, and the applied transform.
}
\description{
Estimate the cross coherence of two signals
}
