% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_estimations_cp.r
\name{tv_estimations_cp}
\alias{tv_estimations_cp}
\title{time-varying estimation of change points}
\usage{
tv_estimations_cp(
  X,
  channelNames = 1:dim(X)[2],
  windowLen,
  M_welch = round_right(windowLen/20),
  overlap = 0,
  normalize = FALSE,
  padEnd = TRUE,
  logScale = FALSE,
  transform = "FZ",
  plot = FALSE,
  check_neighborhood = TRUE,
  check_par = NaN,
  threshFun = function(Tv) {
     0.8 * log(Tv)^(1.1)
 },
  deltaPer = 0.03,
  nCores = 1,
  bands_analysis = FALSE,
  f_sampling = NULL
)
}
\arguments{
\item{X}{numeric array, rows: observations in time, columns: channels.}

\item{channelNames}{list of channel names; if not specified, channels are enumerated.}

\item{windowLen}{numeric, Number of observations used for the estimation of localized spectral quantities.}

\item{M_welch}{numeric, Welch regularisation parameter. Default is windowLen/20}

\item{overlap}{numeric, Number of observations by which localized spectral estimates can overlap. Default is zero.}

\item{normalize}{logical, Normalize spectrum (divide by total localized variance)}

\item{padEnd}{logical, Pad time series with zeros at end (if number of observations in time is not a multiple of windowLen). Default is TRUE}

\item{logScale}{logical, Indicates whether to use a logarithmic scale}

\item{transform}{logical, Transformation of coherence estimate. Fisher-z transform is default and provides consistent change-point estimates.}

\item{plot}{logical, Plot all autospectra and cross-coherences}

\item{check_neighborhood}{logical, Within the change-point detection, should the neighbourhood around a changepoint 
candidate be tested to reduce the risk of spurious detections}

\item{check_par}{numeric, \Delta_T the diameter of the interval around a candidate change-point used in check_neighborhood}

\item{threshFun}{function, Threshold function; see reference for theoretical requirements}

\item{deltaPer}{numeric, Percentage (of total observations in time) of minimum change-point distance.}

\item{nCores}{numeric, the number of cores used in the computation}

\item{bands_analysis}{logical, Should the bands analysis be performed}

\item{f_sampling}{numeric, sampling frequency of the time series}
}
\value{
list of arrays, a list containing the the values of the CUSUM statistics.
}
\description{
time-varying estimation of change points. This function computes the localized spectral quantities; spectrums and cross-coherences,
then applies the cusum_stat function to these quantities.
}
\references{
Shroder & Ombao
}
