% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_spectrum.r
\name{tv_spectrum}
\alias{tv_spectrum}
\title{time-varying spectrum of a univariate time series}
\usage{
tv_spectrum(
  X,
  windowLen,
  overlap = 0,
  M_welch = windowLen/2 * 0.1,
  normalize = FALSE,
  padEnd = TRUE,
  logScale = FALSE,
  bands_analysis = FALSE,
  f_sampling = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{X}{numeric vector, a signal}

\item{windowLen}{numeric, the length of the windows used in the partitin of time domain}

\item{overlap}{numeric, wether the windows overlap or not}

\item{M_welch}{numeric, Welch regularisation parameter. Default is 20% of the windowLen}

\item{normalize}{logical Indicates whether to normalize the spectral quantities.}

\item{padEnd}{logical, Pad time series with zeros at end (if number of observations in time is not a multiple of windowLen). Default is TRUE}

\item{logScale}{logical, Indicates whether to use a logarithmic scale to the estimated periodogram}

\item{plot}{logical Indicates whether to plot the estimated spectrum.}
}
\value{
list, output a list of length 2, the first element is the estimated auto-spectrum, and 
the second, the the set of windows used in the estimation
}
\description{
Estimate the auto-spectrum of a univarite time series
}
