% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHcusum.r
\name{UHcusum}
\alias{UHcusum}
\title{CUSUM statistic}
\usage{
UHcusum(X)
}
\arguments{
\item{X}{Numeric vector, series of numbers}
}
\value{
Numeric vector, cumulative sum using the UH cusum method
}
\description{
Calculate the cumulative sum of a series using the UH cusum method
This function calculates the cumulative sum of a series of numbers
using the UH cusum method.
}
